pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "robokin"
version = "0.1.0"
description = "URDF-based forward and inverse kinematics helpers for robot arms (Placo / PyRoki / RoboPlan)"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Apache-2.0"}
authors = [
    {name = "Dmitri Manajev", email = "dmitri@manajev.com"},
]
maintainers = [
    {name = "Dmitri Manajev", email = "dmitri@manajev.com"},
]
keywords = [
    "robotics",
    "kinematics",
    "inverse-kinematics",
    "urdf",
    "robot-arm",
    "placo",
    "pyroki",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "numpy",
    "scipy",
    "robot_descriptions",
]

[project.urls]
Homepage = "https://github.com/legalaspro/robokin"
Repository = "https://github.com/legalaspro/robokin"
Issues = "https://github.com/legalaspro/robokin/issues"

[project.optional-dependencies]
placo = ["placo"]
pyroki = ["pyroki @ git+https://github.com/chungmin99/pyroki.git"]
# roboplan is not available via pip; install manually via conda-forge / micromamba
# roboplan = ["roboplan"]
viser = ["viser", "yourdfpy"]
rerun = ["rerun-sdk"]
all = [
    "robokin[placo]",
    "robokin[pyroki]",
    "robokin[viser]",
    "robokin[rerun]",
]
dev = [
    "pytest",
    "pytest-cov",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
